backend/requirements.txt
Django>=4.2
psycopg>=3.0
python-dotenv
django-cors-headers
requests
geopandas>=0.14
PyJWT[crypto]>=2.8
PyJWT[crypto]>=2.8
